Server: Add Commit Resolver
