+++ packages/amplication-server/src/core/app/app.resolver.spec.ts
@@ -360,18 +360,6 @@ const GET_COMMITS_QUERY = gql`
   }
 `;
 
-const USER_MUTATION = gql`
-  mutation($message: String!, $appId: String!) {
-    commit(data: { message: $message, app: { connect: { id: $appId } } }) {
-      user {
-        id
-        createdAt
-        updatedAt
-      }
-    }
-  }
-`;
-
 const appMock = jest.fn(() => {
   return EXAMPLE_APP;
 });
@@ -802,30 +790,5 @@ describe('AppResolver', () => {
     });
     expect(getCommitsMock).toBeCalledTimes(1);
     expect(getCommitsMock).toBeCalledWith({});
-  });
-
-  /* Move to independent Commit Resolver Spec if Commit Resolver ever becomes independent */
-  it('should find committing user', async () => {
-    const res = await apolloClient.query({
-      query: USER_MUTATION,
-      variables: {
-        message: EXAMPLE_MESSAGE,
-        appId: EXAMPLE_APP_ID
-      }
-    });
-    expect(res.errors).toBeUndefined();
-    expect(res.data).toEqual({
-      commit: {
-        user: {
-          ...EXAMPLE_USER,
-          createdAt: EXAMPLE_USER.createdAt.toISOString(),
-          updatedAt: EXAMPLE_USER.updatedAt.toISOString()
-        }
-      }
-    });
-    expect(userServiceFindUserMock).toBeCalledTimes(1);
-    expect(userServiceFindUserMock).toBeCalledWith({
-      where: { id: EXAMPLE_USER_ID }
-    });
   });
 });
